backend/users-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.0
python-decouple==3.8
